requirements.txt
anthropic>=3.7.1
openai>=1.6.0
python-dotenv==1.0.0
tenacity==8.2.3